fix: element injection on classical/podcasts, some broken integrations
