+++ src/integration/discord.ts
@@ -2,8 +2,8 @@ import { xxh64 } from "@node-rs/xxhash";
 import { Client } from "@xhayper/discord-rpc";
 import { app } from "electron";
 import log4js, { Logger } from "log4js";
-import { MKPlaybackState } from "~/@types/enums";
-import { PlayerIntegration, TrackMetadata } from "~/@types/interfaces";
+import { MKPlaybackState } from "~/types/enums";
+import { PlayerIntegration, TrackMetadata } from "~/types/interfaces";
 import { AppState } from "~/app/state";
 import SQLiteBackedKV from "~/database/kv";
 import { PlayerSink } from "~/player";
@@ -11,9 +11,14 @@ import { getArtworkUrl, secToMillis, uploadToFreeImageHost } from "~/utils";
 
 const INVISIBLE = "\u00A0";
 
+const MUSIC_APP_ID = "1350945271827136522";
+const CLASSICAL_APP_ID = "1406427068320841788";
+const PODCASTS_APP_ID = "1533912042732519445";
+
 export class DiscordIntegration implements PlayerIntegration {
 	shortName: string = "discord";
 
+	state: AppState;
 	logger: Logger;
 	kvCache: SQLiteBackedKV<string>;
 	playerSink: PlayerSink | null;
@@ -21,18 +26,27 @@ export class DiscordIntegration implements PlayerIntegration {
 	wasPaused: boolean;
 	reconnectTimeout: NodeJS.Timeout | null;
 	constructor(state: AppState) {
+		this.state = state;
 		this.kvCache = new SQLiteBackedKV<string>(app, "discordUrlCache.db");
 		this.logger = log4js.getLogger("discordIntegration");
 		this.logger.level = "debug";
 		this.playerSink = state.playerSink;
 		this.client = new Client({
-			clientId:
-				state.currentWebsite === "music" ?
-					"1350945271827136522"
-				:	"1406427068320841788",
+			clientId: this.getClientId(),
 		});
 		this.wasPaused = false;
 		this.reconnectTimeout = null;
+	}
+
+	private getClientId() {
+		switch (this.state.currentWebsite) {
+			case "classical":
+				return CLASSICAL_APP_ID;
+			case "podcasts":
+				return PODCASTS_APP_ID;
+			default:
+				return MUSIC_APP_ID;
+		}
 	}
 
 	async load() {
@@ -98,14 +112,9 @@ export class DiscordIntegration implements PlayerIntegration {
 		if (originalArtworkUrl && artworkUrl && artworkUrl.length > 256) {
 			try {
 				// remove the AMZ signing params to properly hash
-				const strippedArtworkUrl = originalArtworkUrl.replace(
-					/\?.+/,
-					"",
-				);
+				const strippedArtworkUrl = originalArtworkUrl.replace(/\?.+/, "");
 
-				const hash = xxh64(Buffer.from(strippedArtworkUrl)).toString(
-					16,
-				);
+				const hash = xxh64(Buffer.from(strippedArtworkUrl)).toString(16);
 
 				const alreadyCached = this.kvCache.get(hash);
 
@@ -143,28 +152,28 @@ export class DiscordIntegration implements PlayerIntegration {
 
 		await this.client.user?.setActivity({
 			type: 2, // LISTENING
-			details: metadata["name"].padEnd(2, INVISIBLE),
+			details: metadata["name"]?.padEnd(2, INVISIBLE),
 			...(typeof metadata["url"] === "string" ?
 				{
 					detailsUrl: metadata["url"],
 					largeImageUrl: metadata["url"].split("?")[0],
 				}
 			:	{}),
-			state: metadata["artistName"].padEnd(2, INVISIBLE),
+			state: metadata["artistName"]?.padEnd(2, INVISIBLE),
 			largeImageKey: artworkUrl ? artworkUrl : undefined,
-			largeImageText: metadata["albumName"].padEnd(2, INVISIBLE),
+			largeImageText: metadata["albumName"]?.padEnd(2, INVISIBLE),
 			startTimestamp:
 				this.playerSink?.playbackTime ?
 					Date.now() - secToMillis(this.playerSink.playbackTime)
 				:	undefined,
 			endTimestamp:
-				this.playerSink?.playbackTime ?
-					Date.now()
-					+ (metadata.durationInMillis
-						- secToMillis(this.playerSink.playbackTime))
-				:	undefined,
+				Date.now()
+				+ (metadata.durationInMillis
+					- secToMillis(
+						this.playerSink?.playbackTime ? this.playerSink?.playbackTime : 0,
+					)),
 			instance: false,
-			statusDisplayType: 1, // ACTIVITY_STATE
+			statusDisplayType: this.state.currentWebsite !== "podcasts" ? 1 : 0, // ACTIVITY_STATE
 		});
 	}
