+++ tsconfig.json
@@ -16,8 +16,7 @@
 		"strict": true /* Enable all strict type-checking options. */,
 		"skipLibCheck": true /* Skip type checking all .d.ts files. */,
 		"paths": {
-			"~/*": ["./src/*"],
-			"~~/*": ["./src/extra/*"]
+			"~/*": ["./src/*"]
 		}
 	},
 	"include": ["src/**/*.ts", "src/**/*.tsx"],
